requirements.txt
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorclass==2.2.0
colorlog==5.0.1
confluent-kafka==1.6.1
croniter==1.0.13
dnspython==1.16.0
eventlet==0.31.0
faust==1.10.4
Flask==2.0.1
google-auth==1.29.0
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
kafka-python==1.4.7
kubernetes==12.0.1
MarkupSafe==2.0.1
mode==4.3.2
multidict==5.1.0
mypy-extensions==0.4.3
oauthlib==3.1.0
opentracing==1.3.0
pip-autoremove==0.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
robinhood-aiokafka==1.1.6
rsa==4.7.2
six==1.15.0
terminaltables==3.1.0
typing-extensions==3.10.0.0
urllib3==1.26.4
venusian==1.2.0
websocket-client==0.58.0
Werkzeug==2.0.1
yarl==1.6.3
